requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt
python-dotenv
https://gradio-builds.s3.amazonaws.com/queue-disconnect/v3/gradio-3.4b2-py3-none-any.whl
--extra-index-url https://download.pytorch.org/whl/cu113
diffusers
transformers
scipy
ftfy
nvidia-ml-py3
wget
huggingface_hub
transformers[sentencepiece]
torchvision
fairscale
Pillow
timm
-e git+https://github.com/openai/CLIP.git@main#egg=clip
-e git+https://github.com/pharmapsychotic/BLIP.git@main#egg=blip
datasets
flask
flask_cors
flask_expects_json
gradio
Flask-APScheduler
langdetect
nltk
plotly
tokenizers
sentencepiece
streamlit
powerlaw
numpy
pandas
dataclasses
iso639
python_igraph
PyYAML
scikit_learn
seaborn
streamlit-aggrid
numexpr
scikit-learn
pyarrow
json2html
matplotlib
pyyaml
torch
tqdm
boto3
requests
tensorflow
tensorflow-hub
libsixel-python
gdown
cmake
onnxruntime-gpu
opencv-python-headless
opencv-python
paddlepaddle
paddlehub
pyarmor
pycryptodome
openvino
yapf
onnxruntime
imageio
easydict
facexlib
dominate
onnx
scikit-image
imageio-ffmpeg
imageio
ipython  
psutil  
thop
tensorboard
einops
omegaconf 
pytorch-lightning
kornia
regex
clip-anytorch
lpips
dlib
gdown
git+https://github.com/openai/CLIP.git
wheel
